Separate components from ListTodos
