+++ client/src/types.ts
@@ -3,9 +3,23 @@ export default interface todoType {
   description: string,
   owner: string,
   priority: string,
-  day: string | null,
+  day: string | undefined,
   morning: boolean,
   afternoon: boolean,
   evening: boolean,
   completed: boolean
 }
+
+export type todoCollapsType = {
+  todo_id: string,
+  description: string,
+  owner: string,
+  priority: string,
+  day: string | undefined,
+  morning: boolean,
+  afternoon: boolean,
+  evening: boolean,
+  completed: boolean
+}
+
+export interface ToDoContainer extends Array<todoType> {}
